create a create vote system function to create a vote system
