+++ app/page.tsx
@@ -4,7 +4,6 @@ export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
       <Voting />
-
     </main>
   );
 }
